+++ src/components/common/index.ts
@@ -1,6 +1,5 @@
 export { default as Button } from './Button'
 export { default as ButtonCart } from './ButtonCart'
-export { default as Cart } from './Cart'
 export { default as Header } from './Header'
 export { default as HeaderMobileMenu } from './HeaderMobileMenu'
 export { default as Loader } from './Loader'
